Migrating URL parsing to own file and function (still more cleanup to do but this is much easier to manage index now)
